requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
ollama>=0.1.7
python-docx>=0.8.11
pymupdf>=1.23.0
spacy>=3.6.0
reportlab>=4.0.0
pandas>=2.0.0
nltk>=3.8.0
python-dotenv>=1.0.0
